backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.1.0
torchvision==0.16.0
monai==1.3.0
nibabel==5.1.0
numpy==1.26.0
scikit-learn==1.3.0
pandas==2.1.0
Pillow==10.1.0
